Major Advance: Solved the benchmark memory leak.
The leak was not in library code, it was in the gathering of performace metrics.
With performance code removed the "add one DOM element per animation frame" behaviour is stopped.

Refactored benchmark template

Refresh template base on https://rustwasm.github.io/docs/book/game-of-life/implementing.html
because the code example allows debug symbols.
Now the failure to generate debug symbols can be identfied as a bug in the tool chain related to
extremely long name, when the bugs go away 'wasm-pack build --dev' will eventually work.

Now using the latest version of webpack, copy-webpack-plugin

For now, the fallback is to debug in a non browser environment.