Benchmark is now working.
Now that d3_geo_rs is on the brink of the 1.0.0 release
It is time to make sure everything is in working order.

This upgrade is the first to write to a Path2d object rather than a
context2d object. ( this improved performance.)

We now have a webpack.production.config.js copied from the rotating_50m
example app ... as this is state of the art.

Also some code has been converted from javascript to ts.