Simplification: Dropped polygons struct.
Dropped self from suppliment() and gen().

I did this because it made the code cleaner.

I think this will lead to lead small rises and falls in memory consumption.

In any event it seems to clean up the variability when running benchmark.
The time after 200 samples in now rock solid.